+++ docs/src/examples/components/Icon/States/IconExampleDisabled.shorthand.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { Icon } from 'stardust'
+
+const IconExampleDisabled = () => (
+  <div>
+    <Icon disabled color="red" name="users" size="big" />
+    <Icon disabled color="orange" name="users" size="big" />
+    <Icon disabled color="yellow" name="users" size="big" />
+  </div>
+)
+
+export default IconExampleDisabled